refactor(Command): `ValidateCommand`, `OutputCommand` 패키지 변경
