+++ src/main/java/lotto/service/lotto/LottoService.java
@@ -3,8 +3,8 @@ package lotto.service.lotto;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
-import lotto.command.validate.BonusCommand;
-import lotto.command.validate.LottoCommand;
+import lotto.command.view.validate.BonusCommand;
+import lotto.command.view.validate.LottoCommand;
 import lotto.dto.BonusUserInput;
 import lotto.dto.MatchResult;
 import lotto.dto.MatchResults;
