+++ src/main/java/lotto/view/ConsoleView.java
@@ -2,7 +2,7 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import lotto.command.Command;
-import lotto.command.validate.ValidateCommand;
+import lotto.command.view.validate.ValidateCommand;
 import lotto.dto.UserInput;
 
 /**
